Create README.md
Car Point, a company that deals with the sales of cars, embarked on a marketing campaign that lasted the entire 4th quarter of 2022. The company is interested in knowing the sort of effect the marketing campaign had in their sales, most especially in the year 2023.
A dataset containing the daily transactions of the company from January 2022 to December 2023 have been provided and you as the data analyst have been asked to analyse the data, thereby presenting your findings about the effect of the marketing campaign in the company’s 2023 sales.
Attached in the file above is an Excel file containing the data has been provided. 
Data cleaning, analysis and visualization was done using EXCEL.
A presentation slide discussing the findings from analysis was done using power point and also attached in file above